Using BaseStore to centralize store logic [cis-622]
