+++ ZombieGame/Assets/Scripts/Stores/BaseStore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 211e8123b44d5ae44806d03967546762
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
